feat: crud operations in service and types
