+++ src/cats/dto/change-cat.dto.ts
@@ -1,0 +1,5 @@
+import { CreateCatDto } from "./create-cat.dto";
+
+export class ChangeCatDto extends CreateCatDto {
+    
+}
