+++ src/cats/tests/mock/FakeRepo.ts
@@ -1,14 +1,15 @@
+import { CreateCatDto } from "src/cats/dto/create-cat.dto";
 import { CatsEntity } from "src/cats/entities/cats.entity";
 
 export class CatsRepositoryMock {
-    public create(dto): CatsEntity {
+    public create(dto: CreateCatDto): CatsEntity {
         return {
             id: 'string',
             nick: dto.nick,
             role: dto.role
         }
     }
-    public async save(obj): Promise<CatsEntity> {
-        return obj
+    public async save(entity: CatsEntity): Promise<CatsEntity> {
+        return entity;
     }
 }
